requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--trusted-host devpi.office.novi.systems

asgiref==3.2.7            # via django
django==3.0.5             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
python-magic==0.4.15      # via -r requirements.in
pytz==2019.3              # via django
sqlparse==0.3.1           # via django
